Introduce BaseQueryDAO test base class
